KEYCLOAK-11970 move properties from WebAuthnSecretData to WebAuthnCredentialData
